refector : 1차 리펙터링 ( 함수로 빼기 )
