+++ week3/과제/1to50game/src/utils/useArray.js
@@ -1,0 +1,5 @@
+export const createShuffledArray = (start, end) => {
+  return Array.from({ length: end - start + 1 }, (_, i) => i + start).sort(
+    () => Math.random() - 0.5,
+  );
+};
